refactor(sign in component): organize Sign in page components
organized and refactored the components on Sign In Page to improve the overall structure and
readability of the code
